+++ src/app/login/login.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import LoginPage from './page';
+import Login from './page';
 
 jest.mock('@/assets/icons/ic_Invisibility.svg', () => {
   function MockInVisibilityIcon() {
@@ -17,7 +17,7 @@ jest.mock('@/assets/icons/ic_visibility.svg', () => {
 
 describe('LoginPage', () => {
   test('폼과 버튼이 화면에 나타난다.', () => {
-    render(<LoginPage />);
+    render(<Login />);
     expect(screen.getByLabelText('아이디')).toBeInTheDocument();
     expect(screen.getByLabelText('비밀번호')).toBeInTheDocument();
     expect(screen.getByRole('button', { name: '로그인' })).toBeInTheDocument();
@@ -26,7 +26,7 @@ describe('LoginPage', () => {
 
   test('모든 인풋이 유효하면 로그인버튼이 활성화된다.', async () => {
     const user = userEvent.setup();
-    render(<LoginPage />);
+    render(<Login />);
     await user.type(screen.getByLabelText('아이디'), 'testemail@email.com');
     await user.type(screen.getByLabelText('비밀번호'), 'testpassword');
     expect(screen.getByRole('button', { name: '로그인' })).toBeEnabled();
@@ -34,7 +34,7 @@ describe('LoginPage', () => {
 
   test('이메일이 유효하지 않으면 로그인버튼이 비활성화된다.', async () => {
     const user = userEvent.setup();
-    render(<LoginPage />);
+    render(<Login />);
     await user.type(screen.getByLabelText('아이디'), 'testemail@email');
     await user.type(screen.getByLabelText('비밀번호'), 'testpassword');
     expect(screen.getByRole('button', { name: '로그인' })).toBeDisabled();
@@ -42,7 +42,7 @@ describe('LoginPage', () => {
 
   test('비밀번호가 유효하지 않으면 로그인버튼이 비활성화된다.', async () => {
     const user = userEvent.setup();
-    render(<LoginPage />);
+    render(<Login />);
     await user.type(screen.getByLabelText('아이디'), 'testemail@email.com');
     await user.type(screen.getByLabelText('비밀번호'), 'test');
     expect(screen.getByRole('button', { name: '로그인' })).toBeDisabled();
